build user and team roles lists in a hook
